+++ client/src/pages/Notifications/NotificationsList/useStyles.ts
@@ -1,10 +1,9 @@
 import { makeStyles } from '@material-ui/core/styles';
 
-const useStyles = makeStyles(() => ({
+const useStyles = makeStyles((theme) => ({
   root: {
-    minHeight: '80vh',
-    paddingTop: 40,
-    paddingBottom: 20,
+    minHeight: '60vh',
+    backgroundColor: theme.palette.background.paper,
   },
 
   titleColumn: {
